+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+* Add clear button to Filter area
+* Changed Type of FDTFilter.value
+
 ## 0.0.3
 
 * Negative Page bag fixed.